+++ config/EURUSD_M5.h
@@ -19,10 +19,10 @@ struct Stg_RSI_Params_M5 : StgParams {
     lot_size = 0;
     signal_open_method = 1;
     signal_open_filter = 4;
-    signal_open_level = 30;
+    signal_open_level = (float)30;
     signal_open_boost = 0;
     signal_close_method = 4;
-    signal_close_level = 25;
+    signal_close_level = (float)25;
     price_stop_method = 1;
     price_stop_level = 25;
     tick_filter_method = 1;
